replace custom css with PF, move image detail actions to new component
