+++ src/App.scss
@@ -56,5 +56,4 @@
 
 .remove-list-style {
   list-style-type: none;
-  margin-right: 1.4rem;
 }
