+++ src/Routes/ImageManagerDetail/DetailsHeader.js
@@ -1,54 +1,22 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 import {
-  Text,
   TextList,
   TextListItem,
   TextContent,
   Skeleton,
   Breadcrumb,
   BreadcrumbItem,
-  Button,
   Split,
   SplitItem,
-  Dropdown,
-  DropdownItem,
-  KebabToggle,
 } from '@patternfly/react-core';
 import StatusLabel from './StatusLabel';
+import ImageDetailActions from './ImageDetailActions';
 import { routes as paths } from '../../../package.json';
 
 const DetailsHead = ({ imageData, openUpdateWizard }) => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const dropdownItems = [
-    <DropdownItem
-      listItemClassName="remove-list-style"
-      href={imageData?.Installer?.ImageBuildISOURL}
-      key="link"
-    >
-      <Text className="force-text-black">Download</Text>
-    </DropdownItem>,
-  ];
-
-  const handleToggle = (isOpen) => setIsOpen(isOpen);
-
-  const handleSelect = () => {
-    setIsOpen((prevState) => !prevState);
-    onFocus();
-  };
-
-  const onFocus = () => {
-    const element = document.getElementById('toggle-id-6');
-    element.focus();
-  };
-
-  const handleUpdate = () => {
-    openUpdateWizard(imageData.ID);
-  };
-
   return (
     <>
       <Breadcrumb>
@@ -76,21 +44,10 @@ const DetailsHead = ({ imageData, openUpdateWizard }) => {
           </SplitItem>
           <SplitItem isFilled></SplitItem>
           {imageData?.Status === 'SUCCESS' ? (
-            <SplitItem>
-              <Button onClick={handleUpdate} variant="secondary">
-                Update
-              </Button>
-              <Dropdown
-                position="right"
-                onSelect={handleSelect}
-                toggle={
-                  <KebabToggle onToggle={handleToggle} id="toggle-id-6" />
-                }
-                isOpen={isOpen}
-                isPlain
-                dropdownItems={dropdownItems}
-              />
-            </SplitItem>
+            <ImageDetailActions
+              imageData={imageData}
+              openUpdateWizard={openUpdateWizard}
+            />
           ) : null}
         </Split>
       </TextContent>
